# openapi.yaml
openapi: 3.1.0
info:
  title: DevMate Core API Specification
  description: |
    Production-ready REST and RPC API contract for the DevMate Telegram-based Personal Operating System platform.
    Defines schemas, endpoints, error response wrappers, webhooks, and callbacks for delivery adapters and future clients.
  version: 1.0.0
  termsOfService: https://devmate.internal/terms
  contact:
    name: DevMate API Platform Team
    email: platform-api@devmate.internal
    url: https://devmate.internal/support
  license:
    name: Proprietary License
    url: https://devmate.internal/license
...
servers:
  - url: https://api.devmate.internal/v1
    description: Production Server Gateway
  - url: https://staging-api.devmate.internal/v1
    description: Staging Environment Server
  - url: https://dev-api.devmate.internal/v1
    description: Development Integration Server
  - url: http://localhost:8080/v1
    description: Local Developer Container Server
security:
  - BearerAuth: []
paths:
...
  /notes:
    get:
      summary: List personal notes
      description: Retrieve user note snippets and titles.
      operationId: listNotes
      tags:
        - Notes
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Notes list compiled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: Create note snippet
      description: Logs a new text note.
      operationId: createNote
      tags:
        - Notes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
      responses:
        '201':
          description: Note snippet registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /notes/{id}:
    get:
      summary: Get note details
      description: Retrieves content details and active versions index for a note.
      operationId: getNote
      tags:
        - Notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Note content retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    patch:
      summary: Update note snippet
      description: Modifies text parameters and tags bindings, incrementing version numbers.
      operationId: updateNote
      tags:
        - Notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNoteRequest'
      responses:
        '200':
          description: Note saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete note record
      description: Soft deletes a note snippet.
      operationId: deleteNote
      tags:
        - Notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Note removed successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /finance/budgets:
    get:
      summary: List budgets
      description: Retrieve user budgets status list.
      operationId: listBudgets
      tags:
        - Budgets
      responses:
        '200':
          description: Budgets list compiled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: Define budget limit
      description: Configures alert parameters for category spend limits.
      operationId: createBudget
      tags:
        - Budgets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBudgetRequest'
      responses:
        '201':
          description: Budget limit defined
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /finance/loans:
    get:
      summary: List loans
      description: Retrieve logged active loans.
      operationId: listLoans
      tags:
        - Loans
      responses:
        '200':
          description: Loan records list compiled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: Log active loan
      description: Records debt parameters to auto-generate monthly expense ledgers.
      operationId: createLoan
      tags:
        - Loans
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoanRequest'
      responses:
        '201':
          description: Loan logged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /finance/subscriptions:
    get:
      summary: List billing subscriptions
      description: Retrieve user subscriptions status list.
      operationId: listSubscriptions
      tags:
        - Subscriptions
      responses:
        '200':
          description: Active subscriptions list compiled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: Register subscription billing
      description: Logs billing values to schedule automated repeat transaction alerts.
      operationId: createSubscription
      tags:
        - Subscriptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        '201':
          description: Subscription details logged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /vault/folders:
    get:
      summary: List folders
      description: Retrieve vault folder structures directory lists.
      operationId: listFolders
      tags:
        - Personal Vault
      parameters:
        - name: parentId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Directory catalog compiled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      summary: Create folder
      description: Registers folder directory path in the vault.
      operationId: createFolder
      tags:
        - Personal Vault
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderRequest'
      responses:
        '201':
          description: Folder created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /vault/folders/{id}:
    get:
      summary: Get folder details
      description: Retrieve metadata and files list for folder.
      operationId: getFolder
      tags:
        - Personal Vault
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Folder metadata compiled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    patch:
      summary: Update folder details
      description: Rename folder parameters.
      operationId: updateFolder
      tags:
        - Personal Vault
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 150
                  example: Taxes
      responses:
        '200':
          description: Folder renamed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete folder
      description: Soft deletes directory folder.
      operationId: deleteFolder
      tags:
        - Personal Vault
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Folder deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived session access token passed in Authorization headers.
    RefreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Secure HTTP-only refresh token used to request new access tokens.
    TelegramSecretVerification:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Shared bot secret payload validating inbound Webhook calls.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Reserved API key token for external script integrations.
    OAuth2Placeholder:
      type: oauth2
      description: Reserved for future oauth integrations.
      flows:
        authorizationCode:
          authorizationUrl: https://auth.devmate.internal/oauth/authorize
          tokenUrl: https://auth.devmate.internal/oauth/token
          scopes:
            read: Grant view permissions on owned resources.
            write: Grant modification permissions on owned resources.
  parameters:
...
    LimitParam:
      name: limit
      in: query
      description: Max record items count.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    CursorParam:
      name: cursor
      in: query
      description: Encoded cursor pointer for page offsets.
      required: false
      schema:
        type: string
...
  headers:
    CorrelationIdResponseHeader:
      description: Ingress trace correlation tracker ID UUID.
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Input validation constraints violations.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationIdResponseHeader'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorEnvelope'
...
    Unauthorized:
      description: Authentication failure due to missing or expired tokens.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: The requested resource could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    ServerError:
      description: Internal system error log tracker.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    CreateNoteRequest:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 150
        content:
          type: string
    UpdateNoteRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 150
        content:
          type: string
    NoteResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/NoteDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    NoteDetail:
      type: object
      required:
        - id
        - title
        - content
        - version
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        version:
          type: integer
    NoteListResponse:
      type: object
      required:
        - success
        - data
        - pagination
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/NoteDetail'
        pagination:
          $ref: '#/components/schemas/CursorPaginationMeta'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    CreateBudgetRequest:
      type: object
      required:
        - category
        - limitAmount
      properties:
        category:
          type: string
          minLength: 1
          maxLength: 100
        limitAmount:
          $ref: '#/components/schemas/Money'
...
    BudgetResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/BudgetDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    BudgetDetail:
      type: object
      required:
        - id
        - category
        - limitAmount
        - currentSpend
      properties:
        id:
          type: string
          format: uuid
        category:
          type: string
        limitAmount:
          $ref: '#/components/schemas/Money'
        currentSpend:
          $ref: '#/components/schemas/Money'
    BudgetListResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/BudgetDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    CreateLoanRequest:
      type: object
      required:
        - principal
        - durationMonths
        - startDate
      properties:
        principal:
          $ref: '#/components/schemas/Money'
        interestRate:
          type: string
          pattern: '^\d+(\.\d{1,2})?$'
        durationMonths:
          type: integer
          minimum: 1
        startDate:
          type: string
          format: date
...
    LoanResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/LoanDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    LoanDetail:
      type: object
      required:
        - id
        - principal
        - durationMonths
        - interestRate
        - startDate
        - status
        - version
      properties:
        id:
          type: string
          format: uuid
        principal:
          $ref: '#/components/schemas/Money'
        interestRate:
          type: string
        durationMonths:
          type: integer
        startDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/LoanStatus'
        version:
          type: integer
    LoanListResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/LoanDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    CreateSubscriptionRequest:
      type: object
      required:
        - name
        - amount
        - cycle
        - nextBilling
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 150
        amount:
          $ref: '#/components/schemas/Money'
        cycle:
          type: string
          example: Monthly
        nextBilling:
          type: string
          format: date-time
...
    SubscriptionResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/SubscriptionDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    SubscriptionDetail:
      type: object
      required:
        - id
        - name
        - amount
        - cycle
        - nextBilling
        - status
        - version
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        amount:
          $ref: '#/components/schemas/Money'
        cycle:
          type: string
        nextBilling:
          type: string
          format: date-time
        status:
          type: string
        version:
          type: integer
    SubscriptionListResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/SubscriptionDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    CreateFolderRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 150
        parentId:
          type: string
          format: uuid
    FolderResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/FolderDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    FolderDetail:
      type: object
      required:
        - id
        - name
        - version
      properties:
        id:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: integer
    FolderListResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/FolderDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    Money:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: string
          pattern: '^\d+(\.\d{1,4})?$'
          example: '15.50'
        currency:
          type: string
          maxLength: 3
          minLength: 3
          example: USD
    ResponseMeta:
      type: object
      required:
        - correlationId
        - traceId
      properties:
        correlationId:
          type: string
          format: uuid
        traceId:
          type: string
          format: uuid
    CursorPaginationMeta:
      type: object
      required:
        - hasMore
      properties:
        nextCursor:
          type: string
        hasMore:
          type: boolean
    ErrorEnvelope:
      type: object
      required:
        - success
        - error
        - meta
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required:
            - code
            - message
            - timestamp
          properties:
            code:
              type: string
              example: RESOURCE_NOT_FOUND
            message:
              type: string
              example: Request resource not found.
            timestamp:
              type: string
              format: date-time
            details:
              type: array
              items:
                type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ValidationErrorEnvelope:
      type: object
      required:
        - success
        - error
        - meta
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required:
            - code
            - message
            - timestamp
            - details
          properties:
            code:
              type: string
              const: VALIDATION_FAILED
            message:
              type: string
              example: Invalid input parameters.
            timestamp:
              type: string
              format: date-time
            details:
              type: array
              items:
                type: object
                required:
                  - field
                  - issue
                properties:
                  field:
                    type: string
                    example: amount
                  issue:
                    type: string
                    example: Value must be a positive string.
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    LoanStatus:
      type: string
      enum: [ACTIVE, SETTLED, DEFAULTED]
...